RemoteOscilloscope/pyproject.toml
[project]
name = "remoteoscilloscope"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "jijinbei", email = "hikoharu14142@outlook.jp" }
]
dependencies = [
    "ipykernel>=6.29.5",
    "matplotlib>=3.9.0",
    "polars>=1.0.0",
    "python-dotenv>=1.0.1",
    "pyvisa>=1.14.1",
    "pyvisa-py>=0.7.2",
    "psutil>=6.0.0",
    "zeroconf>=0.132.2",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/remoteoscilloscope"]
